refactor: move checkpoint to lib/infra/
Part of #7697 (restructure src/lib/ into layered architecture).
Move lib/checkpoint/ to lib/infra/checkpoint/ since this module
manages checkpoint/restore — pure infrastructure.

Closes #7724
