+++ turbo/apps/web/src/lib/infra/run/types.ts
@@ -1,4 +1,4 @@
-import type { ArtifactSnapshot } from "../../checkpoint/types";
+import type { ArtifactSnapshot } from "../checkpoint/types";
 import type { Firewalls } from "@vm0/core";
 
 /**
